requirements.txt
# SAPN Exporter dependencies

# HTTP requests and web scraping
requests==2.31.0
beautifulsoup4==4.12.3

# InfluxDB for time-series data storage
influxdb-client==1.38.0

# Task scheduling
APScheduler==3.10.4

# Environment configuration
python-dotenv==1.0.1
